requirements.txt
streamlit
camelot-py[cv]
pandas
